+++ tools/sdmx_import/samples/fetch_eurostat_gdp_data.py
@@ -13,7 +13,7 @@ import os
 sys.path.append(
     os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..')))
 
-from tools.sdmx.dataflow import SdmxClient
+from tools.sdmx_import.sdmx_client import SdmxClient
 
 # Configure logging
 logging.basicConfig(level=logging.INFO,
@@ -46,10 +46,10 @@ def main():
 
     # --- 2. Use the SdmxClient ---
     client = SdmxClient(endpoint, agency_id)
-    client.fetch_and_save_data_as_csv(dataflow_id=dataflow_id,
-                                      key=data_key,
-                                      params=data_params,
-                                      output_path=output_path)
+    client.download_data_as_csv(dataflow_id=dataflow_id,
+                                key=data_key,
+                                params=data_params,
+                                output_path=output_path)
     logging.info(f"--- Successfully downloaded data to {output_path} ---")
